[USBHostMSD] Updated FATFileSystem methods
Updated the virtual FATFileSystem methods for the new argument types

+++ libraries/USBHost/USBHostMSD/USBHostMSD.h
@@ -59,10 +59,10 @@ protected:
     // From FATFileSystem
     virtual int disk_initialize();
     virtual int disk_status() {return 0;};
-    virtual int disk_read(uint8_t* buffer, uint64_t sector, uint8_t count);
-    virtual int disk_write(const uint8_t* buffer, uint64_t sector, uint8_t count);
+    virtual int disk_read(uint8_t* buffer, uint32_t sector, uint32_t count);
+    virtual int disk_write(const uint8_t* buffer, uint32_t sector, uint32_t count);
     virtual int disk_sync() {return 0;};
-    virtual uint64_t disk_sectors();
+    virtual uint32_t disk_sectors();
 
 private:
     USBHost * host;
@@ -104,7 +104,7 @@ private:
     int getMaxLun();
 
     int blockSize;
-    uint64_t blockCount;
+    uint32_t blockCount;
 
     int msd_intf;
     bool msd_device_found;
